refactor(Services): clean code in Gift and Payment

+++ backend/src/main/java/com/capisite/backend/modules/gifts/GiftService.java
@@ -1,7 +1,10 @@
 package com.capisite.backend.modules.gifts;
 
 import java.util.List;
+import java.util.Optional;
 
+import com.capisite.backend.exceptions.ResourceNotFoundException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.capisite.backend.modules.gifts.dto.UpdateGiftDTO;
@@ -17,41 +20,36 @@ public class GiftService {
         this.giftRepository = giftRepository;
     }
 
-    public List<Gift> getAllProducts() {
+    public List<Gift> getAllGifts() {
         return giftRepository.findAll();
     }
 
-    public Gift getProductById(Long id) {
-        return giftRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Produto com ID " + id + " não foi encontrado"));
+    public Gift getGiftById(Long id) {
+        Optional<Gift> gift = giftRepository.findById(id);
+        return gift.orElseThrow(() -> new ResourceNotFoundException(id));
     }
 
-    public Gift createProduct(Gift product) {
-        return giftRepository.save(product);
+    public Gift createGift(Gift gift) {
+        // Criar DTO de criação de Gift
+        return giftRepository.save(gift);
     }
 
-    public Gift updateProduct(Long id, UpdateGiftDTO dto) {
-        Gift product = giftRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Produto com ID " + id + " não foi encontrado"));
-
-        if (dto.name() != null) {
-            product.setName(dto.name());
-        }
-        if (dto.description() != null) {
-            product.setDescription(dto.description());
-        }
-        if (dto.price() != null) {
-            product.setPrice(dto.price());
-        }
-        if (dto.image() != null) {
-            product.setImage(dto.image());
-        }
-
-        return giftRepository.save(product);
+    public Gift updateGift(Long id, UpdateGiftDTO data) {
+        Gift gift = getGiftById(id);
+        updateData(gift, data);
+        return giftRepository.save(gift);
     }
 
-    public void deleteProduct(Long id) {
+    public void deleteGift(Long id) {
+        if (!giftRepository.existsById(id)) throw new ResourceNotFoundException(id);
         giftRepository.deleteById(id);
+    }
+
+    private void updateData(Gift updatedGift, UpdateGiftDTO data) {
+        if(data.name() != null) updatedGift.setName(data.name());
+        if(data.description() != null) updatedGift.setDescription(data.description());
+        if(data.price() != null) updatedGift.setPrice(data.price());
+        if(data.image() != null) updatedGift.setImage(data.image());
     }
 
 }
